requirements.txt
numpy
einops
scipy
pyyaml
tensorboardX
prettytable
timm
nuscenes-devkit
transformers
datasets
tqdm
evaluate
peft
matplotlib
nltk
rouge_score
py7zr
pyblur3